third_party/blink/web_tests/fast/frames/iframe-reparenting-unique-name-expected.txt



--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
This test PASSED if DRT outputs two <iframe>s with unique names.

 

--------
Frame: '<!--framePath //<!--frame1-->/<!--frame0-->-->'
--------
Frame 0

--------
Frame: '<!--framePath //<!--frame1-->/<!--frame1-->-->'
--------
Frame 1
